backend/requirements.txt
numpy==2.2.4
packaging==24.2
pillow==11.1.0
PyYAML==6.0.2
requests==2.32.3
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
numpy==2.2.4
uvicorn==0.34.0
fastapi==0.115.12
python-multipart==0.0.20
openai==1.72.0
python-dotenv==1.1.0
pyarrow==19.0.1
scikit-learn==1.6.1
datasets==3.5.0